+++ src/constants/index.ts
@@ -1,2 +1,4 @@
 export * from './Commands';
 export * from './StateKeys';
+export * from './extensions';
+export * from './languages';