+++ src/models/FileFormat.ts
@@ -1,0 +1,9 @@
+export enum FileFormat {
+  svg = 'svg',
+  png = 'png',
+  jpg = 'jpg',
+  gif = 'gif',
+  bmp = 'bmp',
+  tiff = 'tiff',
+  ico = 'ico',
+}