+++ src/models/index.ts
@@ -1,3 +1,6 @@
 export * from './ContentData';
+export * from './FileCollection';
+export * from './FileFormat';
+export * from './LanguageCollection';
 export * from './Theme';
 export * from './VSCodeMessage';